refactor: 모듈화
#19
