+++ index.html
@@ -75,7 +75,6 @@
         </footer>
     </div>
 
-    <script type="module" src="script.js"></script>
-    <script src="fullscreen.js"></script>
+    <script type="module" src="./js/app.js"></script>
 </body>
 </html>
