+++ js/config.js
@@ -1,0 +1,16 @@
+export const ROWS = 10;
+export const COLS = 20;
+export const INITIAL_TIME_LIMIT = 120;
+export const TARGET_SUM = 11;
+
+export const INITIAL_STATE = {
+  apples: [],
+  selectedApples: [],
+  score: 0,
+  isGameOver: false,
+  timeLimit: INITIAL_TIME_LIMIT,
+  isBGMPlaying: false,
+  isDragging: false,
+  startX: 0,
+  startY: 0
+};
